Fix for 64bit CPUs - use 32bit 'int' instead of 64bit 'long'

+++ ref-impl/include/AAFResult.h
@@ -8,7 +8,7 @@
 
 //=---------------------------------------------------------------------=
 //
-// $Id: AAFResult.h,v 1.62 2004/03/26 20:45:04 creederickson Exp $ $Name:  $
+// $Id: AAFResult.h,v 1.63 2004/10/25 14:01:55 stuart_hc Exp $ $Name:  $
 //
 // The contents of this file are subject to the AAF SDK Public
 // Source License Agreement (the "License"); You may not use this file
@@ -35,7 +35,7 @@
 
 #ifndef MAKE_HRESULT
 #define MAKE_HRESULT(sev,fac,code) \
-    ((HRESULT) (((unsigned long)(sev)<<31) | ((unsigned long)(fac)<<16) | ((unsigned long)(code))) )
+    ((HRESULT) (((unsigned int)(sev)<<31) | ((unsigned int)(fac)<<16) | ((unsigned int)(code))) )
 #endif
 
 // #define _FACILITY_AAF    FACILITY_ITF
